+++ src/Plivo/Resources/MultiPartyCall/MultiPartyCallInterface.php
@@ -353,6 +353,24 @@ class MultiPartyCallInterface extends ResourceInterface
         else{
             $optionalArgs['exit_sound_method'] = 'GET';
         }
+        if(isset($optionalArgs['start_recording_audio'])){
+            MPCUtils::validUrl('startRecordingAudio', $optionalArgs['start_recording_audio'], false);
+        }
+        if(isset($optionalArgs['start_recording_audio_method'])){
+            MPCUtils::validParam('startRecordingAudioMethod', strtoupper($optionalArgs['start_recording_audio_method']), ['string'], false, ['GET', 'POST']);
+        }
+        else{
+            $optionalArgs['start_recording_audio_method'] = 'GET';
+        }
+        if(isset($optionalArgs['stop_recording_audio'])){
+            MPCUtils::validUrl('stopRecordingAudio', $optionalArgs['stop_recording_audio'], false);
+        }
+        if(isset($optionalArgs['stop_recording_audio_method'])){
+            MPCUtils::validParam('stopRecordingAudioMethod', strtoupper($optionalArgs['stop_recording_audio_method']), ['string'], false, ['GET', 'POST']);
+        }
+        else{
+            $optionalArgs['stop_recording_audio_method'] = 'GET';
+        }
         $mandatoryArgs = ['role' => $role];
         $optionalArgs['isVoiceRequest'] = true;
         $response = $this->client->update(
@@ -740,6 +758,58 @@ class MultiPartyCallInterface extends ResourceInterface
         $optionalArgs['isVoiceRequest'] = true;
         $response = $this->client->fetch(
             $this->uri. $mpcId. '/Participant/'. $participantId. '/',
+            $optionalArgs
+        );
+        return $response->getContent();
+    }
+
+    public function startPlayAudio($participantId, $url, array $optionalArgs = []){
+        MPCUtils::validParam('participantId', $participantId, ['string', 'integer'], true);
+        MPCUtils::validUrl('url', $url, true);
+        if(isset($optionalArgs['uuid'])){
+            MPCUtils::validParam('uuid', $optionalArgs['uuid'], ['string'],false);
+        }
+        if(isset($optionalArgs['friendly_name'])){
+            MPCUtils::validParam('friendly_name', $optionalArgs['friendly_name'], ['string'],false);
+        }
+        if(!isset($optionalArgs['uuid'])){
+            $optionalArgs['uuid'] = null;
+        }
+        if(!isset($optionalArgs['friendly_name'])){
+            $optionalArgs['friendly_name'] = null;
+        }
+        $mpcId = self::mpcId($optionalArgs['uuid'], $optionalArgs['friendly_name']);
+        unset($optionalArgs['uuid']);
+        unset($optionalArgs['friendly_name']);
+        $mandatoryArgs = ['url' => $url];
+        $optionalArgs['isVoiceRequest'] = true;
+        $response = $this->client->update(
+            $this->uri. $mpcId. '/Member/'. $participantId. '/Play/',
+            array_merge($mandatoryArgs, $optionalArgs)
+        );
+        return $response->getContent();
+    }
+
+    public function stopPlayAudio($participantId, array $optionalArgs = []){
+        MPCUtils::validParam('participantId', $participantId, ['string', 'integer'], true);
+        if(isset($optionalArgs['uuid'])){
+            MPCUtils::validParam('uuid', $optionalArgs['uuid'], ['string'],false);
+        }
+        if(isset($optionalArgs['friendly_name'])){
+            MPCUtils::validParam('friendly_name', $optionalArgs['friendly_name'], ['string'],false);
+        }
+        if(!isset($optionalArgs['uuid'])){
+            $optionalArgs['uuid'] = null;
+        }
+        if(!isset($optionalArgs['friendly_name'])){
+            $optionalArgs['friendly_name'] = null;
+        }
+        $mpcId = self::mpcId($optionalArgs['uuid'], $optionalArgs['friendly_name']);
+        unset($optionalArgs['uuid']);
+        unset($optionalArgs['friendly_name']);
+        $optionalArgs['isVoiceRequest'] = true;
+        $response = $this->client->delete(
+            $this->uri. $mpcId. '/Member/'. $participantId. '/Play/',
             $optionalArgs
         );
         return $response->getContent();