+++ CHANGELOG.md
@@ -1,5 +1,14 @@
 # Change Log
 
+## [v4.22.0](https://github.com/plivo/plivo-node/tree/v4.22.0) - (2021-11-03)
+- Added Voice MPC enhancements.
+
+## [v4.21.3](https://github.com/plivo/plivo-php/releases/tag/v4.21.3) - 2021-10-29
+- Fix `exception` returned by [Search phone number API](https://www.plivo.com/docs/numbers/api/phone-number/search-phone-numbers/) for successful request.
+
+## [v4.21.2](https://github.com/plivo/plivo-php/releases/tag/v4.21.2) - 2021-10-19
+- Fixing the `GET` request.
+
 ## [v4.21.1](https://github.com/plivo/plivo-php/releases/tag/v4.21.1) - 2021-07-27
 - Updated default HTTP client request timeout to 5 seconds.